requirements.txt
astroid
backports.weakref
bleach
bson
cbor
certifi
chardet
cycler
decorator
entrypoints
h5py
html5lib
idna
influxdb
ipykernel
ipython
ipython-genutils
ipywidgets
isort
jedi
Jinja2
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
Keras
lazy-object-proxy
Markdown
MarkupSafe
matplotlib
mccabe
mistune
msgpack-python
nbconvert
nbformat
networkx
notebook
numpy
olefile
pandas
pandocfilters
pep8
pexpect
pickleshare
Pillow
pkg-resources
plotly
prompt-toolkit
protobuf
ptyprocess
pydotplus
Pygments
pylint
pyparsing
PyQt5
python-dateutil
pytz
PyWavelets
PyYAML
pyzmq
qtconsole
requests
scikit-image
scikit-learn
scipy
seaborn
simplegeneric
simplejson
sip
six
sklearn
tensorflow
tensorflow-gpu
tensorflow-tensorboard
terminado
testpath
Theano
tornado
traitlets
ujson
urllib3
wcwidth
webencodings
Werkzeug
widgetsnbextension
wrapt
xgboost
